feat: display map

+++ InfoTechInterview/Model/City/City.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 struct Coordinates: Decodable {
-	var longitude: Float
-	var latitude: Float
+	var longitude: Double
+	var latitude: Double
 
 	enum CodingKeys: String, CodingKey {
 		case longitude = "lon", latitude = "lat"
